% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{buildQuestion}
\alias{buildQuestion}
\title{Build a question definition from scratch.}
\usage{
buildQuestion(id, dataIds = id, ui, validate = NULL)
}
\arguments{
\item{id}{(character) the unique identifier of the question; it will be used
as the column name in the data.frame returning the questionnaire data
and when prefixed with \code{'question'} - as the \code{inputId}
for the \code{input} slot.}

\item{dataIds}{(character) the list of identifiers that should be
    moved from the \code{input} slot to the final data.frame.

    When a question definition consists of multiple input fields,
    this argument needs to specify the identifiers of those fields.}

\item{ui}{(list|function) a list of UI components of the question or
a function that should return the UI components of the question.
Use a list to build static content and a function to build dynamic content.}

\item{validate}{(function) a function that should return
an empty string if the entered data is valid and a non-empty string,
with the validation message, if the entered data is not valid.}
}
\description{
\code{buildQuestion} returns a question definition given the parts
    a fully formed question definition consists of.
}
\seealso{
\code{\link{buildNonQuestion}},
    \code{\link{question.list}},
    \code{\link{question.numeric}},
    \code{\link{question.text}}.

Other question buidling functions: \code{\link{makeQuestionInputId}}
}

