% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{buildResponses}
\alias{buildResponses}
\title{Build a definition of a response list.}
\usage{
buildResponses(id, label)
}
\arguments{
\item{id}{(character) the list of unique response identifiers.}

\item{label}{(character) the list of unique response labels.}
}
\description{
\code{buildResponses} returns response definitions given
    response identifiers and labels.
}
\seealso{
\code{\link{randomizeResponses}}.

Other response buidling functions: \code{\link{mergeResponses}}
}
