% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{getResponseIds}
\alias{getResponseIds}
\title{Get the identifiers of responses given by the respondent to another question.}
\usage{
getResponseIds(questionId)
}
\arguments{
\item{questionId}{(character) the identifier of the question.}
}
\description{
\code{getResponseIds} returns the identifiers to responses
    mentioned by the respondent in an earlier question.
    Can be \code{NULL} if no response was given.
    In case of text questions, returns the text verbatim.
}

