% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{makeQuestionInputId}
\alias{makeQuestionInputId}
\title{Define a question input identifier.}
\usage{
makeQuestionInputId(id)
}
\arguments{
\item{id}{(character) the unique identifier of the question;
    it will be prefixed with \code{'question'} to create the \code{inputId}
    for the \code{input} slot. Only input fields
    with the \code{'question'} prefix will be returned in the final data.frame.

    For use only when defining questions from scratch. When using question
    builing functions, use the question identifier only.}
}
\description{
\code{makeQuestionInputId} returns a standard question identifier used
for defining input fields.
}
\seealso{
\code{\link{question.list}},
    \code{\link{question.numeric}},
    \code{\link{question.text}}.

Other question buidling functions: \code{\link{buildQuestion}}
}
