% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{maskResponses}
\alias{maskResponses}
\title{Mask the response list using responses given to another question.}
\usage{
maskResponses(responses, questionId, operation = "keep")
}
\arguments{
\item{responses}{(response list) the list of responses
(e.g. as returned by \code{\link{buildResponses}}).}

\item{questionId}{(character) the identifier of the question
that has the reference list of responses.}

\item{operation}{(character) if set to \code{'keep'}, only those responses
in both response lists will be returned; if set to \code{'drop'},
responses given in the reference question will be dropped from
the given response list.}
}
\description{
\code{maskResponses} drops or keeps those responses from the response list
    that were mentioned in an earlier question.
}
\seealso{
\code{\link{buildResponses}}.

Other response handling functions: \code{\link{randomizeResponses}}
}
