% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{mergeResponses}
\alias{mergeResponses}
\title{Merge two or more response lists.}
\usage{
mergeResponses(...)
}
\arguments{
\item{...}{(response lists) the list of response lists
    (e.g. as returned by \code{\link{buildResponses}}).

    Response identifiers and labels must still be unique
    after the merge operation.}
}
\description{
\code{mergeResponses} returns a single response list
    as a merge of individual response lists.
}
\seealso{
\code{\link{randomizeResponses}}.

Other response buidling functions: \code{\link{buildResponses}}
}

