% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{mixedOptions.multiple}
\alias{mixedOptions.multiple}
\title{Get the multiple (non-mutually-exclusive) option type for \code{mixedOptionsInput}.}
\format{An object of class \code{character} of length 1.}
\usage{
mixedOptions.multiple
}
\description{
Returns the multiple (non-mutually-exclusive) option type for \code{mixedOptionsInput}'s \code{types} parameter.
}
\seealso{
\code{\link{mixedOptionsInput}}.

Other mixedOptionsInput option types: \code{\link{mixedOptions.single}}
}
\keyword{datasets}
