% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{mixedOptions.single}
\alias{mixedOptions.single}
\title{Get the single (mutually-exclusive) option type for \code{mixedOptionsInput}.}
\format{An object of class \code{character} of length 1.}
\usage{
mixedOptions.single
}
\description{
Returns the single (mutually-exclusive) option type for \code{mixedOptionsInput}'s \code{types} parameter.
}
\seealso{
\code{\link{mixedOptionsInput}}.

Other mixedOptionsInput option types: \code{\link{mixedOptions.multi}}
}
\keyword{datasets}
