% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{mixedOptionsInput}
\alias{mixedOptionsInput}
\title{Define an input field with both radio-buttons and check-boxes.}
\usage{
mixedOptionsInput(inputId, label, choices, types, selected = NULL,
  inline = FALSE, width = NULL)
}
\arguments{
\item{inputId}{(character) the \code{input} slot that will be used to access the value.}

\item{label}{(character) display label for the control, or \code{NULL} for no label.}

\item{choices}{(list) list of values to show radio-buttons and check-boxes for;
if elements of the list are named then that name, rather than the value,
will be displayed to the user.}

\item{selected}{(vector) the values that should be initially selected, if any.}

\item{inline}{(logical) if \code{TRUE}, render the choices inline (i.e. horizontally).}

\item{width}{(character) the width of the input, e.g. \code{'400px'} or \code{'100\%'};
see \code{\link[shiny]{validateCssUnit}}.}
}
\description{
\code{mixedOptionsInput} creates a set of radio-buttons and check-boxes.
    Check-boxes can be used to toggle multiple choices indepdenently,
    while radio-buttons can be used to define additional, mutually exclusive options.
}
\seealso{
\code{\link{question.mixed}},
    \code{\link[shiny]{checkboxGroupInput}},
    \code{\link[shiny]{radioButtons}}.
}

