% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{question.multiple}
\alias{question.multiple}
\title{Define a multiple-choice question.}
\usage{
question.multiple(id, label, responses, required = TRUE, use.select = FALSE,
  inline = FALSE, width = NULL, placeholder = NULL)
}
\arguments{
\item{id}{(character) the unique identifier of the question; it will be used
as the column name in the data.frame returning the questionnaire data
and when prefixed with \code{'question'} - as the \code{inputId}
for the \code{input} slot.
Note: multiple responses will be coded in a single data.frame column
as comma-separated response identifiers.}

\item{label}{(character) the text displayed as the header of the question.}

\item{responses}{(response list) the response list giving the identifiers and labels
of all responses (e.g. as returned by \code{\link{buildResponses}}).}

\item{required}{(logical) if \code{FALSE}, the respondent is free to not choose
a response; if \code{TRUE}, the respondent must select a response before
moving on to subsequent pages of the questionnaire.}

\item{use.select}{(logical) if \code{FALSE}, displays check-boxes;
if \code{TRUE}, displays a combo-box with all responses available
through the drop-down list.}

\item{inline}{(logical) if \code{FALSE}, check-boxes will be displayed vertically;
if \code{TRUE}, controls will be displayed horizontally.
If \code{use.select == TRUE}, this argument will be ignored.}

\item{width}{(character) the width of the input, e.g. \code{'400px'} or \code{'100\%'}.}

\item{placeholder}{(character) the text that will be displayed
in the combo-box when there are no responses selected yet;
defaults to \code{"Click to select responses"}.
If \code{use.select == FALSE}, this argument will be ignored.}
}
\description{
\code{question.multiple} returns a question definition that displays responses
as check-boxes or as a combo-box, allowing for the selection of multiple responses.
}
\seealso{
\code{\link{buildResponses}},
    \code{\link[shiny]{checkboxGroupInput}},
    \code{\link[shiny]{radioButtons}},
    \code{\link[shiny]{selectInput}}.

Other question definitions: \code{\link{question.mixed}},
  \code{\link{question.numeric}},
  \code{\link{question.single}},
  \code{\link{question.text}}
}
