% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{question.numeric}
\alias{question.numeric}
\title{Define a question that allows the selection of numeric values.}
\usage{
question.numeric(id, label, min, max, step = NA, required = TRUE,
  use.slider = FALSE, width = NULL)
}
\arguments{
\item{id}{(character) the unique identifier of the question; it will be used
as the column name in the data.frame returning the questionnaire data
and when prefixed with \code{'question'} - as the \code{inputId}
for the \code{input} slot.}

\item{label}{(character) the text displayed as the header of the question.}

\item{min}{(numeric) minimum allowed value.}

\item{max}{(numeric) maximum allowed value.}

\item{step}{(numeric) interval to use when stepping between \code{min} and \code{max}.}

\item{required}{(logical) if \code{FALSE}, the respondent is free to not choose
a response; if \code{TRUE}, the respondent must select a response before
moving on to subsequent pages of the questionnaire.
If \code{use.slider == TRUE}, this argument will be ignored.}

\item{use.slider}{(logical) if \code{FALSE}, displays an input line
with a spinner to increase/decrease the value; if \code{TRUE},
displays a slider control.}

\item{width}{(character) the width of the input, e.g. \code{'400px'} or \code{'100\%'}.}
}
\description{
\code{question.numeric} returns a question definition that uses
an input line with a spinner or a slider control for entry of numeric values.
}
\seealso{
\code{\link[shiny]{numericInput}},
    \code{\link[shiny]{sliderInput}}.

Other question definitions: \code{\link{question.mixed}},
  \code{\link{question.multiple}},
  \code{\link{question.single}},
  \code{\link{question.text}}
}
