% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{question.text}
\alias{question.text}
\title{Define a question that allows text entry.}
\usage{
question.text(id, label, required = TRUE, use.textArea = FALSE,
  width = NULL, height = NULL, rows = NULL, placeholder = NULL,
  regex = NULL, regexHint = NULL)
}
\arguments{
\item{id}{(character) the unique identifier of the question; it will be used
as the column name in the data.frame returning the questionnaire data
and when prefixed with 'question' - as the \code{inputId} 
for the \code{input} slot.}

\item{label}{(character) the text displayed as the header of the question.}

\item{required}{(logical) if \code{FALSE}, the respondent is free to not choose
a response; if \code{TRUE}, the respondent must select a response before 
moving on to subsequent pages of the questionnaire.}

\item{use.textArea}{(logical) if \code{FALSE}, displays a single-row edit line;
if \code{TRUE}, displays a larger text area.}

\item{width}{(character) the width of the input, e.g. \code{'400px'} or \code{'100\%'}.}

\item{height}{(character) the height of the input, e.g. \code{'400px'} or \code{'100\%'}.
If \code{use.textArea == FALSE}, this argument is ignored.}

\item{rows}{(character) the height of the input in text rows.
If \code{use.textArea == FALSE}, this argument is ignored.
If \code{height} is set, this argument is ignored.}

\item{placeholder}{(character) the text that will be displayed
in the edit line when no text is entered yet.}

\item{regex}{(character) a regular expression that can be used to validate the entered value.}

\item{regexHint}{(character) a human-readable description of the regular expression
to be displayed when the entered value doesn't match the \code{regex}.
If \code{regex} is not set, this argument is ignored.}
}
\description{
\code{question.text} retuns a question definition that uses
an input line or text area for entry of text values.
}
\seealso{
\code{\link[shiny]{textAreaInput}},
    \code{\link[shiny]{textInput}}.

Other question.definitions: \code{\link{question.list}},
  \code{\link{question.numeric}}
}

