% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{questionnaire}
\alias{questionnaire}
\title{Define an \code{interviewer} questionnaire.}
\usage{
questionnaire(label, welcome, goodbye, exit, ...)
}
\arguments{
\item{label}{(character) the name/label/text that will be displayed
at the top of the UI area assigned to the questionnaire.}

\item{welcome}{(Shiny tag list) the Shiny output that will be displayed
as the first page of the questionnaire - the welcome page.}

\item{goodbye}{(Shiny tag list) the Shiny output that will be displayed
as the last page of the questionnaire, after the respondent
replies to the last question.}

\item{exit}{(function(data.frame)) the callback function that will be called
after the interview finishes; the data.frame will contain responses
to all interview questions; column names will be named using
question ids, and the single row will have the respondent's answers.}

\item{...}{question definitions, non-question UI output,
and functions defining the logic of the questionnaire.}
}
\description{
\code{questionnaire} is the central function via which the whole questionnaire,
    including its questions, responses, and flow logic, is defined.
}
\details{
See \url{https://github.com/mtrybulec/interviewer/blob/master/QUICK-START.md}
   for details on using \code{questionnaire} in Shiny apps.
}
\seealso{
\code{\link{runExample}}.
}

