% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{randomizeResponses}
\alias{randomizeResponses}
\title{Randomize the order of responses in a response list.}
\usage{
randomizeResponses(responses)
}
\arguments{
\item{responses}{(response list) the list of responses
(e.g. as returned by \code{\link{buildResponses}}).}
}
\description{
\code{randomizeResponses} retuns the same response list
    but randomly ordered.
}
\seealso{
\code{\link{buildResponses}},
    \code{\link{mergeResponses}}.
}

