% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{useInterviewer}
\alias{useInterviewer}
\title{Set up \code{interviewer} in a Shiny app.}
\usage{
useInterviewer()
}
\description{
\code{useInterviewer} should be called from a Shiny app's UI
    in order for some of the other \code{interviewer} functions to work correctly.
    It can be called from anywhere inside the UI.
    It also calls \code{useShinyjs()} to set up \code{shinyjs}.
}
\seealso{
\code{\link{runExample}},
    \code{\link[shinyjs]{useShinyjs}}.
}

