% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{buildNonQuestion}
\alias{buildNonQuestion}
\title{Build a non-question (UI) element to show next to questions.}
\usage{
buildNonQuestion(ui)
}
\arguments{
\item{ui}{(list|function) a list of UI components to show on screen
or a function that should return the UI components to show on screen.
Use a list to build static content and a function to build dynamic content.}
}
\description{
\code{buildNonQuestion} returns a non-question definition
    that displays arbitray UI elements.
}
\seealso{
\code{\link{buildQuestion}}.
}
