% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{question.list}
\alias{question.list}
\title{Define a question that displays a list of possible responses.}
\usage{
question.list(id, label, responses, multiple = FALSE, required = TRUE,
  use.select = FALSE, inline = FALSE, width = NULL,
  selectizePlaceholder = NULL)
}
\arguments{
\item{id}{(character) the unique identifier of the question; it will be used
as the column name in the data.frame returning the questionnaire data
and when prefixed with \code{'question'} - as the \code{inputId}
for the \code{input} slot.}

\item{label}{(character) the text displayed as the header of the question.}

\item{responses}{(response list) the response list giving the identifiers and labels
of all responses (e.g. as returned by \code{\link{buildResponses}}).}

\item{multiple}{(logical) if \code{FALSE}, defines a single-choice question;
if \code{TRUE}, dfines a multiple-choice question.}

\item{required}{(logical) if \code{FALSE}, the respondent is free to not choose
a response; if \code{TRUE}, the respondent must select a response before
moving on to subsequent pages of the questionnaire.}

\item{use.select}{(logical) if \code{FALSE}, displays radio-buttons
for single-choice questions and check-boxes for multiple-choice
questions; if \code{TRUE}, displays a combo-box with all responses available
through the drop-down list.}

\item{inline}{(logical) if \code{FALSE}, radio-buttons and check-boxes will be
displayed vertically; if \code{TRUE}, controls will be displayed horizontally.
If \code{use.select == TRUE}, this argument will be ignored.}

\item{width}{(character) the width of the input, e.g. \code{'400px'} or \code{'100\%'}.}

\item{selectizePlaceholder}{(character) the text that will be displayed
in the combo-box when there are no responses selected yet; defaults to
\code{"Click to select a response"} for single-choice questions and
\code{"Click to select responses"} for multiple-choice questions.
If \code{use.select == FALSE}, this argument will be ignored.}
}
\description{
\code{question.list} returns a question definition that displays responses
as radio-buttons, check-boxes, or a combo-box. Can be used for single-
and multiple-choice questions.
}
\seealso{
\code{\link{buildResponses}},
    \code{\link[shiny]{checkboxGroupInput}},
    \code{\link[shiny]{radioButtons}},
    \code{\link[shiny]{selectInput}}.

Other question definitions: \code{\link{question.mixed}},
  \code{\link{question.numeric}},
  \code{\link{question.text}}
}
